+++ class/Personaje.cs
@@ -4,7 +4,7 @@ namespace videoGame
     {
         public Datos DatosPersonaje { get; private set; }
 
-        public Personaje(Datos datos, int velocidad, double destreza, double fuerza, int nivel, double salud)
+        public Personaje(Datos datos, int velocidad, int destreza, double fuerza, int nivel, double salud)
             : base(velocidad, destreza, fuerza, nivel, salud)
         {
             DatosPersonaje = datos;
